refactor(qr-link): move verify into own module, fold into decode feature
Drop the separate `verify` feature — verify_qr and AppAuthenticatedData
are gated behind `decode` instead, since anyone decoding will verify.
Extract verify logic into src/verify.rs to match decode/encode structure.
